# be/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend API for Baper Application",
        "title": "Baper API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/api/auth/login": {
            "post": {
                "description": "Authenticate a user and return JWT tokens",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login User",
                "parameters": [
                    {
                        "description": "Login Credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_auth.AuthResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/auth/register": {
            "post": {
                "description": "Register a new user into the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register New User",
                "parameters": [
                    {
                        "description": "Registration Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_auth.RegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/business/register": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Register a new business for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Business"
                ],
                "summary": "Register a Business",
                "parameters": [
                    {
                        "description": "Business Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_features_busines.RegisterBusinessRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/baper_internal_common_res.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/baper_internal_common_res.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/baper_internal_common_res.Response"
                        }
                    }
                }
            }
        },
...
        "/api/webhook/": {
            "get": {
                "description": "Verify webhook endpoint for Meta WhatsApp API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "Webhook Verification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Mode",
                        "name": "hub.mode",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Challenge",
                        "name": "hub.challenge",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Verify Token",
                        "name": "hub.verify_token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "challenge",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/baper_internal_common_res.Response"
                        }
                    }
                }
            }
        },
        "/api/webhook/send-media": {
            "post": {
                "description": "Send a media message (image/document/video/audio) using a public URL",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "Send Media via URL",
                "parameters": [
                    {
                        "description": "Media Message Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_features_chat.SendMediaMessage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/baper_internal_common_res.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/baper_internal_common_res.Response"
                        }
                    }
                }
            }
        },
        "/api/webhook/send-message": {
            "post": {
                "description": "Send a text message via WhatsApp",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "Send Text Message",
                "parameters": [
                    {
                        "description": "Message Data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_features_chat.SendMessage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/baper_internal_common_res.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/baper_internal_common_res.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/baper_internal_common_res.Response"
                        }
                    }
                }
            }
        },
        "/api/webhook/upload-media": {
            "post": {
                "description": "Upload a local file and send it as a media message",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "Send Media via File Upload",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Destination Phone Number (e.g. 62812...)",
                        "name": "to",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Media Type (image, document, video, audio)",
                        "name": "type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Optional Caption",
                        "name": "caption",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "File to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/baper_internal_common_res.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/baper_internal_common_res.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/baper_internal_common_res.Response"
                        }
                    }
                }
            }
        },
        "/api/webhook/verify": {
            "post": {
                "description": "Receive incoming messages from Meta WhatsApp API",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Webhook"
                ],
                "summary": "Receive Webhook",
                "parameters": [
                    {
                        "description": "Webhook Payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_features_chat.WebhookPayload"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/baper_internal_common_res.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/baper_internal_common_res.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/baper_internal_common_res.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "baper_internal_common_res.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "internal_modules_auth.AuthData": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/internal_modules_auth.UserResponse"
                }
            }
        },
        "internal_modules_auth.AuthResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_modules_auth.AuthData"
                },
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "boolean"
                }
            }
        },
        "internal_modules_auth.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "internal_modules_auth.RegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "first_name",
                "last_name",
                "password",
                "phone"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "minLength": 6
                },
                "phone": {
                    "type": "string"
                }
            }
        },
        "internal_modules_auth.UserResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "internal_modules_features_busines.RegisterBusinessRequest": {
            "type": "object",
            "required": [
                "address",
                "name",
                "phone_business"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone_business": {
                    "type": "string"
                }
            }
        },
        "internal_modules_features_chat.Change": {
            "type": "object",
            "properties": {
                "field": {
                    "type": "string"
                },
                "value": {
                    "$ref": "#/definitions/internal_modules_features_chat.ChangeValue"
                }
            }
        },
        "internal_modules_features_chat.ChangeValue": {
            "type": "object",
            "properties": {
                "contacts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_modules_features_chat.Contact"
                    }
                },
                "messages": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_modules_features_chat.Message"
                    }
                },
                "messaging_product": {
                    "type": "string"
                },
                "metadata": {
                    "$ref": "#/definitions/internal_modules_features_chat.Metadata"
                }
            }
        },
        "internal_modules_features_chat.Contact": {
            "type": "object",
            "properties": {
                "profile": {
                    "$ref": "#/definitions/internal_modules_features_chat.Profile"
                },
                "wa_id": {
                    "type": "string"
                }
            }
        },
        "internal_modules_features_chat.Entry": {
            "type": "object",
            "properties": {
                "changes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_modules_features_chat.Change"
                    }
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "internal_modules_features_chat.Message": {
            "type": "object",
            "properties": {
                "from": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "text": {
                    "$ref": "#/definitions/internal_modules_features_chat.MessageText"
                },
                "timestamp": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "internal_modules_features_chat.MessageText": {
            "type": "object",
            "properties": {
                "body": {
                    "type": "string"
                }
            }
        },
        "internal_modules_features_chat.Metadata": {
            "type": "object",
            "properties": {
                "display_phone_number": {
                    "type": "string"
                },
                "phone_number_id": {
                    "type": "string"
                }
            }
        },
        "internal_modules_features_chat.Profile": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "internal_modules_features_chat.SendMediaMessage": {
            "type": "object",
            "required": [
                "to",
                "type"
            ],
            "properties": {
                "caption": {
                    "type": "string"
                },
                "media_url": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                },
                "type": {
                    "description": "image, document, video, audio",
                    "type": "string"
                }
            }
        },
        "internal_modules_features_chat.SendMessage": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "internal_modules_features_chat.WebhookPayload": {
            "type": "object",
            "properties": {
                "entry": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_modules_features_chat.Entry"
                    }
                },
                "object": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}